+++ src/modules/datastore/datastore_config_plugin.erl
@@ -47,7 +47,6 @@ get_models() -> [
     file_handles,
     sfm_handle,
     custom_metadata,
-    indexes,
     permissions_cache,
     permissions_cache_helper,
     permissions_cache_helper2,
@@ -66,7 +65,8 @@ get_models() -> [
     storage_sync_monitoring,
     storage_sync_info,
     replica_deletion,
-    replica_deletion_lock
+    replica_deletion_lock,
+    index
 ].
 
 %%--------------------------------------------------------------------