+++ test_distributed/replica_eviction_transfers_test_base.erl
@@ -33,7 +33,15 @@
     fail_to_evict_file_replica_without_permissions/3,
     eviction_should_succeed_when_remote_provider_modified_file_replica/3,
     eviction_should_fail_when_evicting_provider_modified_file_replica/3,
-    quota_decreased_after_eviction/3]).
+    quota_decreased_after_eviction/3,
+    schedule_replica_eviction_by_index/2,
+    schedule_eviction_of_regular_file_by_index_with_reduce/2,
+    scheduling_replica_eviction_by_not_existing_index_should_fail/2,
+    scheduling_replica_eviction_by_index_with_wrong_function_should_fail/2,
+    scheduling_replica_eviction_by_empty_index_should_succeed/2,
+    scheduling_replica_eviction_by_not_existing_key_in_index_should_succeed/2,
+    schedule_replica_eviction_of_100_regular_files_by_index/2
+]).
 
 -define(SPACE_ID, <<"space1">>).
 -define(assertQuota(Worker, SpaceId, ExpectedSize), {
@@ -122,6 +130,9 @@ evict_tree_of_empty_dirs(Config, Type, FileKeyType) ->
 
 evict_regular_file_replica(Config, Type, FileKeyType) ->
     [WorkerP2, WorkerP1] = ?config(op_worker_nodes, Config),
+    ProviderId1 = ?GET_DOMAIN_BIN(WorkerP1),
+    ProviderId2 = ?GET_DOMAIN_BIN(WorkerP2),
+    
     transfers_test_mechanism:run_test(
         Config, #transfer_test_spec{
             setup = #setup{
@@ -131,8 +142,8 @@ evict_regular_file_replica(Config, Type, FileKeyType) ->
                 root_directory = transfers_test_utils:root_name(?FUNCTION_NAME, Type, FileKeyType),
                 replicate_to_nodes = [WorkerP2],
                 distribution = [
-                    #{<<"providerId">> => ?GET_DOMAIN_BIN(WorkerP1), <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
-                    #{<<"providerId">> => ?GET_DOMAIN_BIN(WorkerP2), <<"blocks">> => [[0, ?DEFAULT_SIZE]]}
+                    #{<<"providerId">> => ProviderId1, <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
+                    #{<<"providerId">> => ProviderId2, <<"blocks">> => [[0, ?DEFAULT_SIZE]]}
                 ]
             },
             scenario = #scenario{
@@ -155,8 +166,8 @@ evict_regular_file_replica(Config, Type, FileKeyType) ->
                     files_evicted => 1
                 },
                 distribution = [
-                    #{<<"providerId">> => ?GET_DOMAIN_BIN(WorkerP1), <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
-                    #{<<"providerId">> => ?GET_DOMAIN_BIN(WorkerP2), <<"blocks">> => []}
+                    #{<<"providerId">> => ProviderId1, <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
+                    #{<<"providerId">> => ProviderId2, <<"blocks">> => []}
                 ],
                 assertion_nodes = [WorkerP1, WorkerP2]
             }
@@ -165,6 +176,9 @@ evict_regular_file_replica(Config, Type, FileKeyType) ->
 
 evict_regular_file_replica_in_directory(Config, Type, FileKeyType) ->
     [WorkerP2, WorkerP1] = ?config(op_worker_nodes, Config),
+    ProviderId1 = ?GET_DOMAIN_BIN(WorkerP1),
+    ProviderId2 = ?GET_DOMAIN_BIN(WorkerP2),
+    
     transfers_test_mechanism:run_test(
         Config, #transfer_test_spec{
             setup = #setup{
@@ -174,8 +188,8 @@ evict_regular_file_replica_in_directory(Config, Type, FileKeyType) ->
                 root_directory = transfers_test_utils:root_name(?FUNCTION_NAME, Type, FileKeyType),
                 replicate_to_nodes = [WorkerP2],
                 distribution = [
-                    #{<<"providerId">> => ?GET_DOMAIN_BIN(WorkerP1), <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
-                    #{<<"providerId">> => ?GET_DOMAIN_BIN(WorkerP2), <<"blocks">> => [[0, ?DEFAULT_SIZE]]}
+                    #{<<"providerId">> => ProviderId1, <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
+                    #{<<"providerId">> => ProviderId2, <<"blocks">> => [[0, ?DEFAULT_SIZE]]}
                 ]
             },
             scenario = #scenario{
@@ -198,8 +212,8 @@ evict_regular_file_replica_in_directory(Config, Type, FileKeyType) ->
                     files_evicted => 1
                 },
                 distribution = [
-                    #{<<"providerId">> => ?GET_DOMAIN_BIN(WorkerP1), <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
-                    #{<<"providerId">> => ?GET_DOMAIN_BIN(WorkerP2), <<"blocks">> => []}
+                    #{<<"providerId">> => ProviderId1, <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
+                    #{<<"providerId">> => ProviderId2, <<"blocks">> => []}
                 ],
                 assertion_nodes = [WorkerP1, WorkerP2]
             }
@@ -209,6 +223,9 @@ evict_regular_file_replica_in_directory(Config, Type, FileKeyType) ->
 evict_big_file_replica(Config, Type, FileKeyType) ->
     [WorkerP2, WorkerP1] = ?config(op_worker_nodes, Config),
     Size = 1024 * 1024 * 1024,
+    ProviderId1 = ?GET_DOMAIN_BIN(WorkerP1),
+    ProviderId2 = ?GET_DOMAIN_BIN(WorkerP2),
+    
     transfers_test_mechanism:run_test(
         Config, #transfer_test_spec{
             setup = #setup{
@@ -219,8 +236,8 @@ evict_big_file_replica(Config, Type, FileKeyType) ->
                 root_directory = transfers_test_utils:root_name(?FUNCTION_NAME, Type, FileKeyType),
                 replicate_to_nodes = [WorkerP2],
                 distribution = [
-                    #{<<"providerId">> => ?GET_DOMAIN_BIN(WorkerP1), <<"blocks">> => [[0, Size]]},
-                    #{<<"providerId">> => ?GET_DOMAIN_BIN(WorkerP2), <<"blocks">> => [[0, Size]]}
+                    #{<<"providerId">> => ProviderId1, <<"blocks">> => [[0, Size]]},
+                    #{<<"providerId">> => ProviderId2, <<"blocks">> => [[0, Size]]}
                 ]
             },
             scenario = #scenario{
@@ -243,8 +260,8 @@ evict_big_file_replica(Config, Type, FileKeyType) ->
                     files_evicted => 1
                 },
                 distribution = [
-                    #{<<"providerId">> => ?GET_DOMAIN_BIN(WorkerP1), <<"blocks">> => [[0, Size]]},
-                    #{<<"providerId">> => ?GET_DOMAIN_BIN(WorkerP2), <<"blocks">> => []}
+                    #{<<"providerId">> => ProviderId1, <<"blocks">> => [[0, Size]]},
+                    #{<<"providerId">> => ProviderId2, <<"blocks">> => []}
                 ],
                 assertion_nodes = [WorkerP1, WorkerP2]
             }
@@ -253,6 +270,9 @@ evict_big_file_replica(Config, Type, FileKeyType) ->
 
 evict_100_files_in_one_request(Config, Type, FileKeyType) ->
     [WorkerP2, WorkerP1] = ?config(op_worker_nodes, Config),
+    ProviderId1 = ?GET_DOMAIN_BIN(WorkerP1),
+    ProviderId2 = ?GET_DOMAIN_BIN(WorkerP2),
+    
     transfers_test_mechanism:run_test(
         Config, #transfer_test_spec{
             setup = #setup{
@@ -262,8 +282,8 @@ evict_100_files_in_one_request(Config, Type, FileKeyType) ->
                 root_directory = transfers_test_utils:root_name(?FUNCTION_NAME, Type, FileKeyType),
                 replicate_to_nodes = [WorkerP2],
                 distribution = [
-                    #{<<"providerId">> => ?GET_DOMAIN_BIN(WorkerP1), <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
-                    #{<<"providerId">> => ?GET_DOMAIN_BIN(WorkerP2), <<"blocks">> => [[0, ?DEFAULT_SIZE]]}
+                    #{<<"providerId">> => ProviderId1, <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
+                    #{<<"providerId">> => ProviderId2, <<"blocks">> => [[0, ?DEFAULT_SIZE]]}
                 ]
             },
             scenario = #scenario{
@@ -287,8 +307,8 @@ evict_100_files_in_one_request(Config, Type, FileKeyType) ->
                 },
                 assertion_nodes = [WorkerP1, WorkerP2],
                 distribution = [
-                    #{<<"providerId">> => ?GET_DOMAIN_BIN(WorkerP1), <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
-                    #{<<"providerId">> => ?GET_DOMAIN_BIN(WorkerP2), <<"blocks">> => []}
+                    #{<<"providerId">> => ProviderId1, <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
+                    #{<<"providerId">> => ProviderId2, <<"blocks">> => []}
                 ],
                 attempts = 600,
                 timeout = timer:minutes(10)
@@ -298,6 +318,9 @@ evict_100_files_in_one_request(Config, Type, FileKeyType) ->
 
 evict_100_files_each_file_separately(Config, Type, FileKeyType) ->
     [WorkerP2, WorkerP1] = ?config(op_worker_nodes, Config),
+    ProviderId1 = ?GET_DOMAIN_BIN(WorkerP1),
+    ProviderId2 = ?GET_DOMAIN_BIN(WorkerP2),
+    
     transfers_test_mechanism:run_test(
         Config, #transfer_test_spec{
             setup = #setup{
@@ -307,8 +330,8 @@ evict_100_files_each_file_separately(Config, Type, FileKeyType) ->
                 root_directory = transfers_test_utils:root_name(?FUNCTION_NAME, Type, FileKeyType),
                 replicate_to_nodes = [WorkerP2],
                 distribution = [
-                    #{<<"providerId">> => ?GET_DOMAIN_BIN(WorkerP1), <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
-                    #{<<"providerId">> => ?GET_DOMAIN_BIN(WorkerP2), <<"blocks">> => [[0, ?DEFAULT_SIZE]]}
+                    #{<<"providerId">> => ProviderId1, <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
+                    #{<<"providerId">> => ProviderId2, <<"blocks">> => [[0, ?DEFAULT_SIZE]]}
                 ]
             },
             scenario = #scenario{
@@ -332,8 +355,8 @@ evict_100_files_each_file_separately(Config, Type, FileKeyType) ->
                 },
                 assertion_nodes = [WorkerP1, WorkerP2],
                 distribution = [
-                    #{<<"providerId">> => ?GET_DOMAIN_BIN(WorkerP1), <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
-                    #{<<"providerId">> => ?GET_DOMAIN_BIN(WorkerP2), <<"blocks">> => []}
+                    #{<<"providerId">> => ProviderId1, <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
+                    #{<<"providerId">> => ProviderId2, <<"blocks">> => []}
                 ],
                 attempts = 3600,
                 timeout = timer:minutes(60)
@@ -343,6 +366,9 @@ evict_100_files_each_file_separately(Config, Type, FileKeyType) ->
 
 fail_to_evict_file_replica_without_permissions(Config, Type, FileKeyType) ->
     [WorkerP2, WorkerP1] = ?config(op_worker_nodes, Config),
+    ProviderId1 = ?GET_DOMAIN_BIN(WorkerP1),
+    ProviderId2 = ?GET_DOMAIN_BIN(WorkerP2),
+    
     transfers_test_mechanism:run_test(
         Config, #transfer_test_spec{
             setup = #setup{
@@ -352,8 +378,8 @@ fail_to_evict_file_replica_without_permissions(Config, Type, FileKeyType) ->
                 root_directory = transfers_test_utils:root_name(?FUNCTION_NAME, Type, FileKeyType),
                 replicate_to_nodes = [WorkerP2],
                 distribution = [
-                    #{<<"providerId">> => ?GET_DOMAIN_BIN(WorkerP1), <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
-                    #{<<"providerId">> => ?GET_DOMAIN_BIN(WorkerP2), <<"blocks">> => [[0, ?DEFAULT_SIZE]]}
+                    #{<<"providerId">> => ProviderId1, <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
+                    #{<<"providerId">> => ProviderId2, <<"blocks">> => [[0, ?DEFAULT_SIZE]]}
                 ]
             },
             scenario = #scenario{
@@ -367,8 +393,8 @@ fail_to_evict_file_replica_without_permissions(Config, Type, FileKeyType) ->
             expected = #expected{
                 expected_transfer = undefined,
                 distribution = [
-                    #{<<"providerId">> => ?GET_DOMAIN_BIN(WorkerP1), <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
-                    #{<<"providerId">> => ?GET_DOMAIN_BIN(WorkerP2), <<"blocks">> => [[0, ?DEFAULT_SIZE]]}
+                    #{<<"providerId">> => ProviderId1, <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
+                    #{<<"providerId">> => ProviderId2, <<"blocks">> => [[0, ?DEFAULT_SIZE]]}
                 ],
                 assertion_nodes = [WorkerP1, WorkerP2]
             }
@@ -377,6 +403,9 @@ fail_to_evict_file_replica_without_permissions(Config, Type, FileKeyType) ->
 
 eviction_should_succeed_when_remote_provider_modified_file_replica(Config, Type, FileKeyType) ->
     [WorkerP2, WorkerP1] = ?config(op_worker_nodes, Config),
+    ProviderId1 = ?GET_DOMAIN_BIN(WorkerP1),
+    ProviderId2 = ?GET_DOMAIN_BIN(WorkerP2),
+    
     Config2 = transfers_test_mechanism:run_test(
         Config, #transfer_test_spec{
             setup = #setup{
@@ -386,8 +415,8 @@ eviction_should_succeed_when_remote_provider_modified_file_replica(Config, Type,
                 root_directory = transfers_test_utils:root_name(?FUNCTION_NAME, Type, FileKeyType),
                 replicate_to_nodes = [WorkerP2],
                 distribution = [
-                    #{<<"providerId">> => ?GET_DOMAIN_BIN(WorkerP1), <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
-                    #{<<"providerId">> => ?GET_DOMAIN_BIN(WorkerP2), <<"blocks">> => [[0, ?DEFAULT_SIZE]]}
+                    #{<<"providerId">> => ProviderId1, <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
+                    #{<<"providerId">> => ProviderId2, <<"blocks">> => [[0, ?DEFAULT_SIZE]]}
                 ]
             },
             scenario = undefined,
@@ -425,8 +454,8 @@ eviction_should_succeed_when_remote_provider_modified_file_replica(Config, Type,
                     files_evicted => 1
                 },
                 distribution = [
-                    #{<<"providerId">> => ?GET_DOMAIN_BIN(WorkerP1), <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
-                    #{<<"providerId">> => ?GET_DOMAIN_BIN(WorkerP2), <<"blocks">> => []}
+                    #{<<"providerId">> => ProviderId1, <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
+                    #{<<"providerId">> => ProviderId2, <<"blocks">> => []}
                 ],
                 assertion_nodes = [WorkerP1, WorkerP2]
             }
@@ -435,6 +464,9 @@ eviction_should_succeed_when_remote_provider_modified_file_replica(Config, Type,
 
 eviction_should_fail_when_evicting_provider_modified_file_replica(Config, Type, FileKeyType) ->
     [WorkerP2, WorkerP1] = ?config(op_worker_nodes, Config),
+    ProviderId1 = ?GET_DOMAIN_BIN(WorkerP1),
+    ProviderId2 = ?GET_DOMAIN_BIN(WorkerP2),
+    
     Config2 = transfers_test_mechanism:run_test(
         Config, #transfer_test_spec{
             setup = #setup{
@@ -444,8 +476,8 @@ eviction_should_fail_when_evicting_provider_modified_file_replica(Config, Type, 
                 root_directory = transfers_test_utils:root_name(?FUNCTION_NAME, Type, FileKeyType),
                 replicate_to_nodes = [WorkerP2],
                 distribution = [
-                    #{<<"providerId">> => ?GET_DOMAIN_BIN(WorkerP1), <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
-                    #{<<"providerId">> => ?GET_DOMAIN_BIN(WorkerP2), <<"blocks">> => [[0, ?DEFAULT_SIZE]]}
+                    #{<<"providerId">> => ProviderId1, <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
+                    #{<<"providerId">> => ProviderId2, <<"blocks">> => [[0, ?DEFAULT_SIZE]]}
                 ]
             },
             scenario = undefined,
@@ -489,8 +521,8 @@ eviction_should_fail_when_evicting_provider_modified_file_replica(Config, Type, 
                     files_evicted => 0
                 },
                 distribution = [
-                    #{<<"providerId">> => ?GET_DOMAIN_BIN(WorkerP1), <<"blocks">> => [[0,1], [2, ?DEFAULT_SIZE - 2]]},
-                    #{<<"providerId">> => ?GET_DOMAIN_BIN(WorkerP2), <<"blocks">> => [[0, ?DEFAULT_SIZE]]}
+                    #{<<"providerId">> => ProviderId1, <<"blocks">> => [[0,1], [2, ?DEFAULT_SIZE - 2]]},
+                    #{<<"providerId">> => ProviderId2, <<"blocks">> => [[0, ?DEFAULT_SIZE]]}
                 ],
                 assertion_nodes = [WorkerP1, WorkerP2]
             }
@@ -500,7 +532,9 @@ eviction_should_fail_when_evicting_provider_modified_file_replica(Config, Type, 
 quota_decreased_after_eviction(Config, Type, FileKeyType) ->
     [WorkerP2, WorkerP1] = ?config(op_worker_nodes, Config),
     SpaceId = ?config(?SPACE_ID_KEY, Config),
-    ct:pal("SpaceId: ~p", [SpaceId]),
+    ProviderId1 = ?GET_DOMAIN_BIN(WorkerP1),
+    ProviderId2 = ?GET_DOMAIN_BIN(WorkerP2),
+    
     Config2 = transfers_test_mechanism:run_test(
         Config, #transfer_test_spec{
             setup = #setup{
@@ -510,8 +544,8 @@ quota_decreased_after_eviction(Config, Type, FileKeyType) ->
                 root_directory = transfers_test_utils:root_name(?FUNCTION_NAME, Type, FileKeyType),
                 replicate_to_nodes = [WorkerP2],
                 distribution = [
-                    #{<<"providerId">> => ?GET_DOMAIN_BIN(WorkerP1), <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
-                    #{<<"providerId">> => ?GET_DOMAIN_BIN(WorkerP2), <<"blocks">> => [[0, ?DEFAULT_SIZE]]}
+                    #{<<"providerId">> => ProviderId1, <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
+                    #{<<"providerId">> => ProviderId2, <<"blocks">> => [[0, ?DEFAULT_SIZE]]}
                 ]
             },
             scenario = undefined,
@@ -543,8 +577,8 @@ quota_decreased_after_eviction(Config, Type, FileKeyType) ->
                     files_evicted => 1
                 },
                 distribution = [
-                    #{<<"providerId">> => ?GET_DOMAIN_BIN(WorkerP1), <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
-                    #{<<"providerId">> => ?GET_DOMAIN_BIN(WorkerP2), <<"blocks">> => []}
+                    #{<<"providerId">> => ProviderId1, <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
+                    #{<<"providerId">> => ProviderId2, <<"blocks">> => []}
                 ],
                 assertion_nodes = [WorkerP1, WorkerP2],
                 attempts = 120
@@ -552,6 +586,489 @@ quota_decreased_after_eviction(Config, Type, FileKeyType) ->
         }
     ),
     ?assertQuota(WorkerP2, SpaceId, 0).
+
+schedule_replica_eviction_by_index(Config, Type) ->
+    [WorkerP2, WorkerP1] = ?config(op_worker_nodes, Config),
+    SessionId2 = ?DEFAULT_SESSION(WorkerP2, Config),
+    SpaceId = ?SPACE_ID,
+    ProviderId1 = ?GET_DOMAIN_BIN(WorkerP1),
+    ProviderId2 = ?GET_DOMAIN_BIN(WorkerP2),
+    
+    Config2 = transfers_test_mechanism:run_test(
+        Config, #transfer_test_spec{
+            setup = #setup{
+                setup_node = WorkerP1,
+                files_structure = [{0, 1}],
+                root_directory = transfers_test_utils:root_name(?FUNCTION_NAME, Type),
+                replicate_to_nodes = [WorkerP2],
+                distribution = [
+                    #{<<"providerId">> => ProviderId1, <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
+                    #{<<"providerId">> => ProviderId2, <<"blocks">> => [[0, ?DEFAULT_SIZE]]}
+                ],
+                assertion_nodes = [WorkerP1, WorkerP2]
+            }}),
+
+    [{FileGuid, _}] = ?config(?FILES_KEY, Config2),
+    {ok, FileId} = cdmi_id:guid_to_objectid(FileGuid),
+
+    % set xattr on file to be replicated
+    XattrName = transfers_test_utils:random_job_name(),
+    XattrValue = 1,
+    Xattr = #xattr{name = XattrName, value = XattrValue},
+    ok = lfm_proxy:set_xattr(WorkerP2, SessionId2, {guid, FileGuid}, Xattr),
+    IndexName = <<"replica_eviction_jobs">>,
+    MapFunction = transfers_test_utils:test_map_function(XattrName),
+    transfers_test_utils:create_index(WorkerP2, SpaceId, IndexName, MapFunction, [], [ProviderId2]),
+    ?assertIndexQuery([FileId], WorkerP2, SpaceId, IndexName,  [{key, XattrValue}]),
+    ?assertIndexVisible(WorkerP1, SpaceId, IndexName),
+
+    transfers_test_mechanism:run_test(
+        Config2, #transfer_test_spec{
+            setup = undefined,
+            scenario = #scenario{
+                type = Type,
+                schedule_node = WorkerP1,
+                evicting_nodes = [WorkerP2],
+                space_id = SpaceId,
+                function = fun transfers_test_mechanism:evict_replicas_from_index/2,
+                query_view_params = [{key, XattrValue}],
+                index_name = IndexName
+            },
+            expected = #expected{
+                expected_transfer = #{
+                    eviction_status => completed,
+                    scheduling_provider => transfers_test_utils:provider_id(WorkerP1),
+                    evicting_provider => transfers_test_utils:provider_id(WorkerP2),
+                    files_to_process => 2,
+                    files_processed => 2,
+                    files_replicated => 0,
+                    bytes_replicated => 0,
+                    files_evicted =>  1
+                },
+                assertion_nodes = [WorkerP1, WorkerP2],
+                distribution = [
+                    #{<<"providerId">> => ProviderId1, <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
+                    #{<<"providerId">> => ProviderId2, <<"blocks">> => []}
+                ],
+                assert_transferred_file_model = false
+            }
+        }
+    ).
+
+schedule_eviction_of_regular_file_by_index_with_reduce(Config, Type) ->
+    [WorkerP2, WorkerP1] = ?config(op_worker_nodes, Config),
+    SessionId2 = ?DEFAULT_SESSION(WorkerP2, Config),
+    SpaceId = ?SPACE_ID,
+    ProviderId1 = ?GET_DOMAIN_BIN(WorkerP1),
+    ProviderId2 = ?GET_DOMAIN_BIN(WorkerP2),
+
+    Config2 = transfers_test_mechanism:run_test(
+        Config, #transfer_test_spec{
+            setup = #setup{
+                setup_node = WorkerP1,
+                files_structure = [{0, 6}],
+                replicate_to_nodes = [WorkerP2],
+                root_directory = transfers_test_utils:root_name(?FUNCTION_NAME, Type),
+                distribution = [
+                    #{<<"providerId">> => ProviderId1, <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
+                    #{<<"providerId">> => ProviderId2, <<"blocks">> => [[0, ?DEFAULT_SIZE]]}
+                ],
+                assertion_nodes = [WorkerP1, WorkerP2]
+            }}),
+
+
+    FileGuidsAndPaths = ?config(?FILES_KEY, Config2),
+    [Guid1, Guid2, Guid3, Guid4, _Guid5, Guid6 | _] = [G || {G, _} <- FileGuidsAndPaths],
+
+    % XattrName1 will be used by map function
+    % only files with XattrName1 = XattrValue11 will be emitted
+    % XattrName2 will be used by reduce function
+    % only files with XattrName2 = XattrValue12 will be filtered
+    XattrName1 = transfers_test_utils:random_job_name(),
+    XattrName2 = transfers_test_utils:random_job_name(),
+    XattrValue11 = 1,
+    XattrValue12 = 2,
+    XattrValue21 = 1,
+    XattrValue22 = 2,
+    Xattr11 = #xattr{name = XattrName1, value = XattrValue11},
+    Xattr12 = #xattr{name = XattrName1, value = XattrValue12},
+    Xattr21 = #xattr{name = XattrName2, value = XattrValue21},
+    Xattr22 = #xattr{name = XattrName2, value = XattrValue22},
+
+    % File1: xattr1=1, xattr2=1, should be replicated
+    ok = lfm_proxy:set_xattr(WorkerP2, SessionId2, {guid, Guid1}, Xattr11),
+    ok = lfm_proxy:set_xattr(WorkerP2, SessionId2, {guid, Guid1}, Xattr21),
+
+    % File2: xattr1=1, xattr2=2, should not be replicated
+    ok = lfm_proxy:set_xattr(WorkerP2, SessionId2, {guid, Guid2}, Xattr11),
+    ok = lfm_proxy:set_xattr(WorkerP2, SessionId2, {guid, Guid2}, Xattr22),
+
+    % File3: xattr1=1, xattr2=null, should not be replicated
+    ok = lfm_proxy:set_xattr(WorkerP2, SessionId2, {guid, Guid3}, Xattr11),
+
+    % File4: xattr1=2, xattr2=null, should not be replicated
+    ok = lfm_proxy:set_xattr(WorkerP2, SessionId2, {guid, Guid4}, Xattr12),
+
+    % File5: xattr1=null, xattr2=null, should not be replicated
+
+    % File6: xattr1=1, xattr2=1, should be replicated
+    ok = lfm_proxy:set_xattr(WorkerP2, SessionId2, {guid, Guid6}, Xattr11),
+    ok = lfm_proxy:set_xattr(WorkerP2, SessionId2, {guid, Guid6}, Xattr21),
+
+    IndexName = <<"replication_jobs">>,
+    MapFunction = transfers_test_utils:test_map_function(XattrName1, XattrName2),
+    ReduceFunction = transfers_test_utils:test_reduce_function(XattrValue21),
+
+    ok = transfers_test_utils:create_index(WorkerP2, SpaceId, IndexName,
+        MapFunction, ReduceFunction, [{group, 1}, {key, XattrValue11}],
+        [ProviderId2]
+    ),
+
+    {ok, ObjectId1} = cdmi_id:guid_to_objectid(Guid1),
+    {ok, ObjectId6} = cdmi_id:guid_to_objectid(Guid6),
+
+    ?assertIndexQuery([ObjectId1, ObjectId6], WorkerP2, SpaceId, IndexName,  [{key, XattrValue11}]),
+    ?assertIndexVisible(WorkerP1, SpaceId, IndexName),
+
+    transfers_test_mechanism:run_test(
+        Config2, #transfer_test_spec{
+            setup = undefined,
+            scenario = #scenario{
+                type = Type,
+                schedule_node = WorkerP1,
+                evicting_nodes = [WorkerP2],
+                space_id = SpaceId,
+                function = fun transfers_test_mechanism:evict_replicas_from_index/2,
+                query_view_params = [{group, true}, {key, XattrValue11}],
+                index_name = IndexName
+            },
+            expected = #expected{
+                expected_transfer = #{
+                    replication_status => skipped,
+                    scheduling_provider => transfers_test_utils:provider_id(WorkerP1),
+                    evicting_provider => transfers_test_utils:provider_id(WorkerP2),
+                    files_to_process => 3,
+                    files_processed => 3,
+                    files_replicated => 0,
+                    bytes_replicated => 0,
+                    files_evicted => 2
+                },
+                distribution = [
+                    #{<<"providerId">> => ProviderId1, <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
+                    #{<<"providerId">> => ProviderId2, <<"blocks">> => []}
+                ],
+                assert_distribution_for_files = [Guid1, Guid6],
+                assertion_nodes = [WorkerP1, WorkerP2],
+                assert_transferred_file_model = false
+            }
+        }
+    ).
+
+scheduling_replica_eviction_by_not_existing_index_should_fail(Config, Type) ->
+    [WorkerP2, WorkerP1] = ?config(op_worker_nodes, Config),
+    SessionId2 = ?DEFAULT_SESSION(WorkerP2, Config),
+    SpaceId = ?SPACE_ID,
+    ProviderId1 = ?GET_DOMAIN_BIN(WorkerP1),
+    ProviderId2 = ?GET_DOMAIN_BIN(WorkerP2),
+    
+    Config2 = transfers_test_mechanism:run_test(
+        Config, #transfer_test_spec{
+            setup = #setup{
+                setup_node = WorkerP1,
+                files_structure = [{0, 1}],
+                root_directory = transfers_test_utils:root_name(?FUNCTION_NAME, Type),
+                replicate_to_nodes = [WorkerP2],
+                distribution = [
+                    #{<<"providerId">> => ProviderId1, <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
+                    #{<<"providerId">> => ProviderId2, <<"blocks">> => [[0, ?DEFAULT_SIZE]]}
+                ],
+                assertion_nodes = [WorkerP1, WorkerP2]
+            }}),
+
+    [{FileGuid, _}] = ?config(?FILES_KEY, Config2),
+
+    % set xattr on file to be replicated
+    XattrName = transfers_test_utils:random_job_name(),
+    XattrValue = 1,
+    Xattr = #xattr{name = XattrName, value = XattrValue},
+    ok = lfm_proxy:set_xattr(WorkerP2, SessionId2, {guid, FileGuid}, Xattr),
+    IndexName = <<"not_existing_index">>,
+
+    transfers_test_mechanism:run_test(
+        Config2, #transfer_test_spec{
+            setup = undefined,
+            scenario = #scenario{
+                type = Type,
+                schedule_node = WorkerP1,
+                evicting_nodes = [WorkerP2],
+                space_id = SpaceId,
+                function = fun transfers_test_mechanism:fail_to_evict_replicas_from_index/2,
+                query_view_params = [{key, XattrValue}],
+                index_name = IndexName
+            },
+            expected = #expected{
+                assertion_nodes = [WorkerP1, WorkerP2],
+                distribution = [
+                    #{<<"providerId">> => ProviderId1, <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
+                    #{<<"providerId">> => ProviderId2, <<"blocks">> => [[0, ?DEFAULT_SIZE]]}
+                ],
+                assert_transferred_file_model = false
+            }
+        }
+    ).
+
+scheduling_replica_eviction_by_index_with_wrong_function_should_fail(Config, Type) ->
+    [WorkerP2, WorkerP1] = ?config(op_worker_nodes, Config),
+    SessionId2 = ?DEFAULT_SESSION(WorkerP2, Config),
+    SpaceId = ?SPACE_ID,
+    ProviderId1 = ?GET_DOMAIN_BIN(WorkerP1),
+    ProviderId2 = ?GET_DOMAIN_BIN(WorkerP2),
+
+    Config2 = transfers_test_mechanism:run_test(
+        Config, #transfer_test_spec{
+            setup = #setup{
+                setup_node = WorkerP1,
+                files_structure = [{0, 1}],
+                root_directory = transfers_test_utils:root_name(?FUNCTION_NAME, Type),
+                replicate_to_nodes = [WorkerP2],
+                distribution = [
+                    #{<<"providerId">> => ProviderId1, <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
+                    #{<<"providerId">> => ProviderId2, <<"blocks">> => [[0, ?DEFAULT_SIZE]]}
+                ],
+                assertion_nodes = [WorkerP1, WorkerP2]
+            }}),
+
+    [{FileGuid, _}] = ?config(?FILES_KEY, Config2),
+
+    % set xattr on file to be replicated
+    XattrName = transfers_test_utils:random_job_name(),
+    XattrValue = 1,
+    Xattr = #xattr{name = XattrName, value = XattrValue},
+    ok = lfm_proxy:set_xattr(WorkerP2, SessionId2, {guid, FileGuid}, Xattr),
+    WrongValue = <<"random_value_instead_of_file_id">>,
+    %functions does not emit file id in values
+    MapFunction = <<
+        "function (id, meta) {
+            if(meta['", XattrName/binary,"']) {
+                return [meta['", XattrName/binary, "'], '", WrongValue/binary, "'];
+            }
+        return null;
+    }">>,
+    IndexName = <<"index_with_wrong_function">>,
+    transfers_test_utils:create_index(WorkerP2, SpaceId, IndexName, MapFunction, [], [ProviderId2]),
+    ?assertIndexQuery([WrongValue], WorkerP2, SpaceId, IndexName,  [{key, XattrValue}]),
+    ?assertIndexVisible(WorkerP1, SpaceId, IndexName),
+
+    transfers_test_mechanism:run_test(
+        Config2, #transfer_test_spec{
+            setup = undefined,
+            scenario = #scenario{
+                type = Type,
+                schedule_node = WorkerP1,
+                evicting_nodes = [WorkerP2],
+                space_id = SpaceId,
+                function = fun transfers_test_mechanism:evict_replicas_from_index/2,
+                query_view_params = [{key, XattrValue}],
+                index_name = IndexName
+            },
+            expected = #expected{
+                expected_transfer = #{
+                    eviction_status => failed,
+                    scheduling_provider => transfers_test_utils:provider_id(WorkerP1),
+                    evicting_provider => transfers_test_utils:provider_id(WorkerP2),
+                    files_to_process => 2,
+                    files_processed => 2,
+                    files_evicted => 0,
+                    failed_files => 1
+                },
+                assertion_nodes = [WorkerP1, WorkerP2],
+                distribution = [
+                    #{<<"providerId">> => ProviderId1, <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
+                    #{<<"providerId">> => ProviderId2, <<"blocks">> => [[0, ?DEFAULT_SIZE]]}
+                ],
+                assert_transferred_file_model = false
+            }
+        }
+    ).
+
+scheduling_replica_eviction_by_empty_index_should_succeed(Config, Type) ->
+    [WorkerP2, WorkerP1] = ?config(op_worker_nodes, Config),
+    SpaceId = ?SPACE_ID,
+    XattrName = transfers_test_utils:random_job_name(),
+    IndexName = <<"replica_eviction_jobs">>,
+    MapFunction = transfers_test_utils:test_map_function(XattrName),
+    ProviderId2 = ?GET_DOMAIN_BIN(WorkerP2),
+    
+    transfers_test_utils:create_index(WorkerP2, SpaceId, IndexName, MapFunction, [], [ProviderId2]),
+    ?assertIndexQuery([], WorkerP2, SpaceId, IndexName,  []),
+    ?assertIndexVisible(WorkerP1, SpaceId, IndexName),
+
+    transfers_test_mechanism:run_test(
+        Config, #transfer_test_spec{
+            setup = undefined,
+            scenario = #scenario{
+                type = Type,
+                schedule_node = WorkerP1,
+                evicting_nodes = [WorkerP2],
+                space_id = SpaceId,
+                function = fun transfers_test_mechanism:evict_replicas_from_index/2,
+                query_view_params = [],
+                index_name = IndexName
+            },
+            expected = #expected{
+                expected_transfer = #{
+                    eviction_status => completed,
+                    scheduling_provider => transfers_test_utils:provider_id(WorkerP1),
+                    evicting_provider => transfers_test_utils:provider_id(WorkerP2),
+                    files_to_process => 1,
+                    files_processed => 1,
+                    files_evicted => 0
+                },
+                assertion_nodes = [WorkerP1, WorkerP2],
+                distribution = undefined,
+                assert_transferred_file_model = false
+            }
+        }
+    ).
+
+scheduling_replica_eviction_by_not_existing_key_in_index_should_succeed(Config, Type) ->
+    [WorkerP2, WorkerP1] = ?config(op_worker_nodes, Config),
+    SessionId2 = ?DEFAULT_SESSION(WorkerP2, Config),
+    SpaceId = ?SPACE_ID,
+    ProviderId1 = ?GET_DOMAIN_BIN(WorkerP1),
+    ProviderId2 = ?GET_DOMAIN_BIN(WorkerP2),
+    
+    Config2 = transfers_test_mechanism:run_test(
+        Config, #transfer_test_spec{
+            setup = #setup{
+                setup_node = WorkerP1,
+                files_structure = [{0, 1}],
+                root_directory = transfers_test_utils:root_name(?FUNCTION_NAME, Type),
+                replicate_to_nodes = [WorkerP2],
+                distribution = [
+                    #{<<"providerId">> => ProviderId1, <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
+                    #{<<"providerId">> => ProviderId2, <<"blocks">> => [[0, ?DEFAULT_SIZE]]}
+                ],
+                assertion_nodes = [WorkerP1, WorkerP2]
+            }}),
+
+    [{FileGuid, _}] = ?config(?FILES_KEY, Config2),
+    {ok, FileId} = cdmi_id:guid_to_objectid(FileGuid),
+
+    % set xattr on file to be replicated
+    XattrName = transfers_test_utils:random_job_name(),
+    XattrValue = 1,
+    XattrValue2 = 2,
+    Xattr = #xattr{name = XattrName, value = XattrValue},
+    ok = lfm_proxy:set_xattr(WorkerP2, SessionId2, {guid, FileGuid}, Xattr),
+    IndexName = <<"replica_eviction_jobs">>,
+    MapFunction = transfers_test_utils:test_map_function(XattrName),
+    transfers_test_utils:create_index(WorkerP2, SpaceId, IndexName, MapFunction, [], [ProviderId2]),
+    ?assertIndexQuery([FileId], WorkerP2, SpaceId, IndexName,  [{key, XattrValue}]),
+    ?assertIndexVisible(WorkerP1, SpaceId, IndexName),
+
+    transfers_test_mechanism:run_test(
+        Config2, #transfer_test_spec{
+            setup = undefined,
+            scenario = #scenario{
+                type = Type,
+                schedule_node = WorkerP1,
+                evicting_nodes = [WorkerP2],
+                space_id = SpaceId,
+                function = fun transfers_test_mechanism:evict_replicas_from_index/2,
+                query_view_params = [{key, XattrValue2}],
+                index_name = IndexName
+            },
+            expected = #expected{
+                expected_transfer = #{
+                    eviction_status => completed,
+                    scheduling_provider => transfers_test_utils:provider_id(WorkerP1),
+                    evicting_provider => transfers_test_utils:provider_id(WorkerP2),
+                    files_to_process => 1,
+                    files_processed => 1,
+                    files_evicted => 0
+                },
+                assertion_nodes = [WorkerP1, WorkerP2],
+                distribution = [
+                    #{<<"providerId">> => ProviderId1, <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
+                    #{<<"providerId">> => ProviderId2, <<"blocks">> => [[0, ?DEFAULT_SIZE]]}
+                ],
+                assert_transferred_file_model = false
+            }
+        }
+    ).
+
+schedule_replica_eviction_of_100_regular_files_by_index(Config, Type) ->
+    [WorkerP2, WorkerP1] = ?config(op_worker_nodes, Config),
+    SessionId2 = ?DEFAULT_SESSION(WorkerP2, Config),
+    SpaceId = ?SPACE_ID,
+    NumberOfFiles = 100,
+    ProviderId1 = ?GET_DOMAIN_BIN(WorkerP1),
+    ProviderId2 = ?GET_DOMAIN_BIN(WorkerP2),
+    
+    Config2 = transfers_test_mechanism:run_test(
+        Config, #transfer_test_spec{
+            setup = #setup{
+                setup_node = WorkerP1,
+                files_structure = [{0, NumberOfFiles}],
+                root_directory = transfers_test_utils:root_name(?FUNCTION_NAME, Type),
+                replicate_to_nodes = [WorkerP2],
+                distribution = [
+                    #{<<"providerId">> => ProviderId1, <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
+                    #{<<"providerId">> => ProviderId2, <<"blocks">> => [[0, ?DEFAULT_SIZE]]}
+                ],
+                assertion_nodes = [WorkerP1, WorkerP2]
+            }}),
+
+    FileGuidsAndPaths = ?config(?FILES_KEY, Config2),
+
+    % set xattr on file to be replicated
+    XattrName = transfers_test_utils:random_job_name(),
+    XattrValue = 1,
+    Xattr = #xattr{name = XattrName, value = XattrValue},
+
+    FileIds = lists:map(fun({FileGuid, _}) ->
+        ok = lfm_proxy:set_xattr(WorkerP2, SessionId2, {guid, FileGuid}, Xattr),
+        {ok, FileId} = cdmi_id:guid_to_objectid(FileGuid),
+        FileId
+    end, FileGuidsAndPaths),
+
+    IndexName = <<"replica_eviction_jobs">>,
+    MapFunction = transfers_test_utils:test_map_function(XattrName),
+    transfers_test_utils:create_index(WorkerP2, SpaceId, IndexName, MapFunction, [], [ProviderId2]),
+    ?assertIndexQuery(FileIds, WorkerP2, SpaceId, IndexName, [{key, XattrValue}]),
+    ?assertIndexVisible(WorkerP1, SpaceId, IndexName),
+
+    transfers_test_mechanism:run_test(
+        Config2, #transfer_test_spec{
+            setup = undefined,
+            scenario = #scenario{
+                type = Type,
+                schedule_node = WorkerP1,
+                evicting_nodes = [WorkerP2],
+                space_id = SpaceId,
+                function = fun transfers_test_mechanism:evict_replicas_from_index/2,
+                query_view_params = [{key, XattrValue}],
+                index_name = IndexName
+            },
+            expected = #expected{
+                expected_transfer = #{
+                    eviction_status => completed,
+                    scheduling_provider => transfers_test_utils:provider_id(WorkerP1),
+                    evicting_provider => transfers_test_utils:provider_id(WorkerP2),
+                    files_to_process => NumberOfFiles + 1,
+                    files_processed => NumberOfFiles + 1,
+                    files_evicted => NumberOfFiles
+                },
+                assertion_nodes = [WorkerP1, WorkerP2],
+                distribution = [
+                    #{<<"providerId">> => ProviderId1, <<"blocks">> => [[0, ?DEFAULT_SIZE]]},
+                    #{<<"providerId">> => ProviderId2, <<"blocks">> => []}
+                ],
+                assert_transferred_file_model = false
+            }
+        }
+    ).
 
 %%%===================================================================
 %%% SetUp and TearDown functions
@@ -583,6 +1100,18 @@ init_per_suite(Config) ->
 init_per_testcase(quota_decreased_after_eviction, Config) ->
     init_per_testcase(all, [{?SPACE_ID_KEY, <<"space3">>} | Config]);
 
+init_per_testcase(schedule_replica_eviction_of_100_regular_files_by_index_with_batch_100, Config) ->
+    Nodes = [WorkerP2 | _] = ?config(op_worker_nodes, Config),
+    {ok, OldValue} = test_utils:get_env(WorkerP2, op_worker, replica_eviction_by_index_batch),
+    test_utils:set_env(Nodes, op_worker, replica_eviction_by_index_batch, 100),
+    init_per_testcase(all, [{replica_eviction_by_index_batch, OldValue} | Config]);
+
+init_per_testcase(schedule_replica_eviction_of_100_regular_files_by_index_with_batch_10, Config) ->
+    Nodes = [WorkerP2 | _] = ?config(op_worker_nodes, Config),
+    {ok, OldValue} = test_utils:get_env(WorkerP2, op_worker, replica_eviction_by_index_batch),
+    test_utils:set_env(Nodes, op_worker, replica_eviction_by_index_batch, 10),
+    init_per_testcase(all, [{replica_eviction_by_index_batch, OldValue} | Config]);
+
 init_per_testcase(_Case, Config) ->
     ct:timetrap(timer:minutes(60)),
     lfm_proxy:init(Config),
@@ -598,12 +1127,22 @@ end_per_testcase(eviction_should_fail_when_evicting_provider_modified_file_repli
     test_utils:mock_unload(WorkerP2, replica_deletion_req),
     end_per_testcase(all, Config);
 
+end_per_testcase(Case, Config) when
+    Case =:= schedule_replica_eviction_of_100_regular_files_by_index_with_batch_100;
+    Case =:= schedule_replica_eviction_of_100_regular_files_by_index_with_batch_10
+    ->
+    Nodes = ?config(op_worker_nodes, Config),
+    OldValue = ?config(replica_eviction_by_index_batch, Config),
+    test_utils:set_env(Nodes, op_worker, replica_eviction_by_index_batch, OldValue),
+    end_per_testcase(all, Config);
+
 end_per_testcase(_Case, Config) ->
     Workers = ?config(op_worker_nodes, Config),
     transfers_test_utils:unmock_sync_req(Workers),
     transfers_test_utils:unmock_replica_synchronizer_failure(Workers),
     transfers_test_utils:remove_transfers(Config),
     rpc:multicall(Workers, transfer, restart_pools, []),
+    transfers_test_utils:remove_all_indexes(Workers, ?DEFAULT_USER),
     transfers_test_utils:ensure_transfers_removed(Config).
 
 end_per_suite(Config) ->