+++ test_distributed/replication_transfers_lfm_test_SUITE.erl
@@ -45,8 +45,16 @@
     schedule_replication_on_not_supporting_provider_by_guid/1,
     schedule_replication_on_not_supporting_provider_by_path/1,
     file_replication_failures_should_fail_whole_transfer/1,
-    many_simultaneous_failed_transfers/1
-]).
+    many_simultaneous_failed_transfers/1,
+    schedule_replication_of_regular_file_by_index/1,
+    schedule_replication_of_regular_file_by_index_with_reduce/1,
+    scheduling_replication_by_not_existing_index_should_fail/1,
+    scheduling_replication_by_index_with_wrong_function_should_fail/1,
+    scheduling_replication_by_empty_index_should_succeed/1,
+    scheduling_replication_by_not_existing_key_in_index_should_succeed/1,
+    schedule_replication_of_100_regular_files_by_index_with_batch_1000/1,
+    schedule_replication_of_100_regular_files_by_index_with_batch_100/1,
+    schedule_replication_of_100_regular_files_by_index_with_batch_10/1]).
 
 all() -> [
     replicate_empty_dir_by_guid,
@@ -72,7 +80,16 @@ all() -> [
     schedule_replication_on_not_supporting_provider_by_guid,
     schedule_replication_on_not_supporting_provider_by_path,
     % file_replication_failures_should_fail_whole_transfer, TODO uncomment after resolving VFS-4742
-    many_simultaneous_failed_transfers
+    many_simultaneous_failed_transfers,
+    schedule_replication_of_regular_file_by_index,
+    schedule_replication_of_regular_file_by_index_with_reduce,
+    scheduling_replication_by_not_existing_index_should_fail,
+    scheduling_replication_by_index_with_wrong_function_should_fail,
+    scheduling_replication_by_empty_index_should_succeed,
+    scheduling_replication_by_not_existing_key_in_index_should_succeed,
+    schedule_replication_of_100_regular_files_by_index_with_batch_1000,
+    schedule_replication_of_100_regular_files_by_index_with_batch_100,
+    schedule_replication_of_100_regular_files_by_index_with_batch_10
 ].
 
 %%%===================================================================
@@ -151,6 +168,36 @@ file_replication_failures_should_fail_whole_transfer(Config) ->
 
 many_simultaneous_failed_transfers(Config) ->
     replication_transfers_test_base:many_simultaneous_failed_transfers(Config, lfm, guid).
+
+schedule_replication_of_regular_file_by_index(Config) ->
+    replication_transfers_test_base:schedule_replication_of_regular_file_by_index(Config, lfm).
+
+schedule_replication_of_regular_file_by_index_with_reduce(Config) ->
+    replication_transfers_test_base:schedule_replication_of_regular_file_by_index_with_reduce(Config, lfm).
+
+scheduling_replication_by_not_existing_index_should_fail(Config) ->
+    replication_transfers_test_base:scheduling_replication_by_not_existing_index_should_fail(Config, lfm).
+
+scheduling_replication_by_index_with_wrong_function_should_fail(Config) ->
+    replication_transfers_test_base:scheduling_replication_by_index_with_wrong_function_should_fail(Config, lfm).
+
+scheduling_replication_by_empty_index_should_succeed(Config) ->
+    replication_transfers_test_base:scheduling_replication_by_empty_index_should_succeed(Config, lfm).
+
+scheduling_replication_by_not_existing_key_in_index_should_succeed(Config) ->
+    replication_transfers_test_base:scheduling_replication_by_not_existing_key_in_index_should_succeed(Config, lfm).
+
+schedule_replication_of_100_regular_files_by_index_with_batch_1000(Config) ->
+    replication_transfers_test_base:schedule_replication_of_100_regular_files_by_index(Config, lfm).
+
+schedule_replication_of_100_regular_files_by_index_with_batch_100(Config) ->
+    %replication_transfers_test_base:init_per_testcase sets replica_eviction_by_index_batch variable to 100
+    replication_transfers_test_base:schedule_replication_of_100_regular_files_by_index(Config, lfm).
+
+schedule_replication_of_100_regular_files_by_index_with_batch_10(Config) ->
+    %replication_transfers_test_base:init_per_testcase sets replica_eviction_by_index_batch variable to 10
+    replication_transfers_test_base:schedule_replication_of_100_regular_files_by_index(Config, lfm).
+
 
 %%%===================================================================
 %%% SetUp and TearDown functions